2017_11_23customerMoment

+++ database/migrations/2017_09_09_072844_create_suppliers_table.php
@@ -16,7 +16,7 @@ class CreateSuppliersTable extends Migration
         Schema::create('suppliers', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('price_factor');
+            $table->float('price_factor');
             $table->string('code');
             $table->timestamps();
         });
